Methods for csv
This methods are helpers to open a file dialog for csv and another two are for reading purposes. Both other methods read a column of a csv file and store it in a list. Difference between them is, that one method chooses the column by name (string) and the other one by number.